+++ CHANGES.rst
@@ -25,6 +25,7 @@ Changes
    and test builds)
  - shortcut postgresql distribution of builds with the 'not-used'
    special version filter
+ - new pg_remote_copy.timeout option, for the db subfactory
 
 0.9 (2015-05-15)
 ----------------